+++ docs/tailwind.config.cjs
@@ -4,6 +4,7 @@ const defaultTheme = require('tailwindcss/defaultTheme');
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: ['./src/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue}'],
+  darkMode: 'class',
   theme: {
     extend: {
       fontFamily: {
@@ -29,6 +30,9 @@ module.exports = {
             },
             'blockquote p:last-of-type::after': {
               content: '',
+            },
+            'blockquote code': {
+              color: '#FFA657',
             },
           },
         },