+++ examples/graphql/prisma/seed.ts
@@ -1,0 +1,18 @@
+import { PrismaClient } from '@prisma/client';
+
+const prisma = new PrismaClient();
+
+async function main() {
+  await prisma.user.create({
+    data: {
+      name: 'Ham Burger',
+      email: 'ham@burger.dev',
+    },
+  });
+}
+
+main()
+  .catch((e) => console.error(e))
+  .finally(async () => {
+    await prisma.$disconnect();
+  });