+++ examples/graphql/src/app.module.ts
@@ -1,0 +1,22 @@
+import { PrismaModule } from 'nestjs-prisma';
+import { Module } from '@nestjs/common';
+import { GraphQLModule } from '@nestjs/graphql';
+import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
+import { join } from 'path';
+import { UserModule } from './user/user.module';
+
+@Module({
+  imports: [
+    GraphQLModule.forRoot<ApolloDriverConfig>({
+      driver: ApolloDriver,
+      autoSchemaFile: join(process.cwd(), 'src/schema.gql'),
+      sortSchema: true,
+    }),
+    PrismaModule.forRoot({ isGlobal: true }),
+
+    UserModule,
+  ],
+  controllers: [],
+  providers: [],
+})
+export class AppModule {}