+++ examples/graphql/src/user/dto/update-user.input.ts
@@ -1,0 +1,8 @@
+import { CreateUserInput } from './create-user.input';
+import { Field, InputType, PartialType } from '@nestjs/graphql';
+
+@InputType()
+export class UpdateUserInput extends PartialType(CreateUserInput) {
+  @Field(() => String)
+  id: string;
+}